refactor: 유저 정보 조회 관련 code 수정
