+++ src/main/java/com/ninedocs/userserver/user/presentation/usermanagement/UserProfileController.java
@@ -3,7 +3,6 @@ package com.ninedocs.userserver.user.presentation.usermanagement;
 import com.ninedocs.userserver.common.presentation.dto.ApiResponse;
 import com.ninedocs.userserver.user.application.userprofile.UserProfileService;
 import com.ninedocs.userserver.user.application.userprofile.dto.UserProfileResponse;
-import com.ninedocs.userserver.user.persistence.UserRepository;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
@@ -19,13 +18,11 @@ import org.springframework.web.bind.annotation.RestController;
 public class UserProfileController {
 
   private final UserProfileService userProfileService;
-  private final UserRepository userRepository;
 
   @Operation(summary = "유저 정보 조회")
   @GetMapping("/api/v1/user/{userId}")
   public ResponseEntity<ApiResponse<UserProfileResponse>> checkUser(
-      @PathVariable Long userId
-  ) {
+      @PathVariable("userId") long userId) {
     return ResponseEntity.status(HttpStatus.OK).body(ApiResponse.success(
         userProfileService.getUserProfile(userId)
     ));
